[AQUMV] Refactor codes to new file aqumv.c
Refactor to external func answer_query_using_materialized_views
with some refinements.
Rename GUC to enable_answer_query_using_materialized_views to
avoid duplicated name with implementation function.
New function aqumv_init_context() to init context of rewrite.

Authored-by: Zhang Mingli avamingli@gmail.com
